requirements.txt
# Auto-generated requirements for this agent workspace

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread

# Market Research
google-search-results  # SerpAPI

# Data Processing
pandas
openpyxl  # Excel file parsing and creation
xlsxwriter  # Advanced Excel formatting
python-docx  # Word document creation

# Formatting utilities
gspread-formatting  # Google Sheets formatting

# LLM APIs (optional - only for standalone mode)
openai
anthropic

# Utilities
python-dotenv
